Refactor auto-destruct to use OpenProcess for PID monitoring on Windows
