+++ screenpipe-app-tauri/src-tauri/src/sidecar.rs
@@ -104,34 +104,6 @@ pub async fn kill_all_sreenpipes(
     }
 }
 
-#[cfg(target_os = "windows")]
-pub async fn auto_destruct_monitor(auto_destruct_pid: Option<u32>) {
-    if let Some(pid) = auto_destruct_pid {
-        loop {
-            if !is_process_alive(pid) {
-                tracing::info!("Parent process with PID {} is not alive. Terminating sidecar.", pid);
-                process::exit(0);
-            }
-            sleep(Duration::from_secs(1)).await;
-        }
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn is_process_alive(pid: u32) -> bool {
-    use windows::Win32::System::Threading::{OpenProcess, PROCESS_QUERY_INFORMATION};
-    use windows::Win32::Foundation::{CloseHandle, HANDLE};
-
-    unsafe {
-        let process: HANDLE = OpenProcess(PROCESS_QUERY_INFORMATION, false, pid);
-        if process.is_invalid() {
-            return false;
-        }
-        CloseHandle(process);
-        true
-    }
-}
-
 #[tauri::command]
 pub async fn spawn_screenpipe(
     state: tauri::State<'_, SidecarState>,
